WIP on combining stats and search
