+++ src/Search/Search.tsx
@@ -1,20 +1,25 @@
 import { Box, Flex } from '@rebass/grid';
-import fileDownload from 'js-file-download';
 import * as React from 'react';
 
-import { Search } from '../api';
-import { IHacker, ISearchParameter, UserType } from '../config';
-import { Button, H1 } from '../shared/Elements';
+import { Hacker, Search } from '../api';
+import { IHacker, ISearchParameter, IStats, UserType } from '../config';
 import ValidationErrorGenerator from '../shared/Form/validationErrorGenerator';
 import WithToasterContainer from '../shared/HOC/withToaster';
-import theme from '../shared/Styles/theme';
 import { getNestedAttr } from '../util';
 import { FilterComponent } from './Filters';
 import { ResultsTable } from './ResultsTable';
+import { StatsComponent } from './Stats/Stats';
+
+enum SearchMode {
+  TABLE,
+  STATS,
+}
+
 interface ISearchState {
   model: string;
+  mode: SearchMode;
   query: ISearchParameter[];
-  results: Array<{
+  dataTableResults: Array<{
     /**
      * For now, we aren't exposing 'selected' attribute. We set it default equal to true.
      * This is used for batch operations for changing hacker data.
@@ -22,6 +27,7 @@ interface ISearchState {
     selected: boolean;
     hacker: IHacker;
   }>;
+  statsResults: IStats | null;
   loading: boolean;
 }
 
@@ -30,13 +36,15 @@ class SearchContainer extends React.Component<{}, ISearchState> {
     super(props);
     this.state = {
       model: 'hacker',
+      mode: SearchMode.STATS,
       query: [],
-      results: [],
+      dataTableResults: [],
       loading: false,
+      statsResults: null,
     };
     this.onFilterChange = this.onFilterChange.bind(this);
     this.triggerSearch = this.triggerSearch.bind(this);
-    this.downloadData = this.downloadData.bind(this);
+    this.triggerSearch();
   }
   public render() {
     return (
@@ -48,68 +56,38 @@ class SearchContainer extends React.Component<{}, ISearchState> {
           />
         </Box>
         <Box width={5 / 6}>
-          <Flex justifyContent={'space-between'}>
-            <Box>
-              <H1
-                color={theme.colors.primary}
-                fontSize={'30px'}
-                textAlign={'left'}
-                marginTop={'0px'}
-                marginBottom={'20px'}
-              >
-                Hackers
-              </H1>
-            </Box>
-            <Box mr={'10px'}>
-              <Button>Update Status</Button>
-              <Button onClick={this.downloadData}>Export Hackers</Button>
-            </Box>
-          </Flex>
-          <ResultsTable
-            results={this.state.results}
-            loading={this.state.loading}
-            userType={UserType.STAFF}
-          />
+          {this.state.mode === SearchMode.TABLE ? (
+            <ResultsTable
+              results={this.state.dataTableResults}
+              loading={this.state.loading}
+              userType={UserType.STAFF}
+            />
+          ) : (
+            <StatsComponent
+              stats={this.state.statsResults}
+              loading={this.state.loading}
+            />
+          )}
         </Box>
       </Flex>
     );
   }
-
-  private downloadData(): void {
-    const headers = [
-      '_id',
-      'accountId.firstName',
-      'accountId.lastName',
-      'accountId.email',
-      'needsBus',
-      'major',
-      'school',
-      'graduationYear',
-      'degree',
-      'gender',
-      'needsBus',
-    ];
-    const csvData: string[] = [headers.join('\t')];
-    this.state.results.forEach((result) => {
-      if (result.selected) {
-        const row: string[] = [];
-        headers.forEach((header) => {
-          let value;
-          if (header.indexOf('.') >= 0) {
-            const nestedAttr = header.split('.');
-            value = getNestedAttr(result.hacker, nestedAttr);
-          } else {
-            value = result.hacker[header];
-          }
-          row.push(value);
-        });
-        csvData.push(row.join('\t'));
-      }
-    });
-    fileDownload(csvData.join('\n'), 'hackerData.tsv', 'text/tsv');
+  private async triggerStatsSearch(): Promise<void> {
+    try {
+      const statsResponse = await Hacker.getStats(this.state.query);
+      const stats: IStats | null = getNestedAttr(statsResponse, [
+        'data',
+        'data',
+        'stats',
+      ]);
+      this.setState({ statsResults: stats, loading: false });
+    } catch (e) {
+      console.log('Error while getting stats', e);
+      this.setState({ loading: false });
+    }
   }
 
-  private async triggerSearch(): Promise<void> {
+  private async triggerTableSearch(): Promise<void> {
     const { model, query } = this.state;
     try {
       const response = await Search.search(model, query, {
@@ -123,10 +101,20 @@ class SearchContainer extends React.Component<{}, ISearchState> {
             hacker: v,
           }))
         : [];
-      this.setState({ results: tableData, loading: false });
+      this.setState({ dataTableResults: tableData, loading: false });
     } catch (e) {
       ValidationErrorGenerator(e.data);
       this.setState({ loading: false });
+    }
+  }
+
+  private async triggerSearch(): Promise<void> {
+    const { mode } = this.state;
+    switch (mode) {
+      case SearchMode.STATS:
+        return this.triggerStatsSearch();
+      case SearchMode.TABLE:
+        return this.triggerTableSearch();
     }
   }
   private onFilterChange(newFilters: ISearchParameter[]) {
