+++ src/Search/Stats/ActiveShape.tsx
